requirements.txt
flask
joblib
pillow
numpy